+++ src/main/java/uy/edu/ude/BuscadorProyectos/entity/Categoria.java
@@ -22,7 +22,7 @@ public class Categoria {
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	@Column(name="IdCategoria")
-	private Long id;
+	private int id;
 	
 	@NotNull
 	@Size(min = 1, max = 255)
@@ -40,15 +40,15 @@ public class Categoria {
 		this.tecnologias = tecnologias;
 	}
 
-	public Categoria(long idCategoria) {
+	public Categoria(int idCategoria) {
 		this.id = idCategoria;
 	}
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
